+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#1587] The Mesh Settings gizmo was not shown when in `SetOrInherit` mode
 - [#1589] A `KeyNotFoundException` could occur when the target of a Merge Animator or Merge Motion component was null
 - [#1605] Fixed an issue where the preview differed from the build result when multiple Material Setters conflicted
+- [#1632] BlendshapeSync would not work in the editor when on a disabled object
 
 ### Changed
 - [#1608] [#1610] Shape Changer will now delete shapekeys fully, even if animated